pyproject.toml
[project]
name = "api-with-flask"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-restful (>=0.3.10,<0.4.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)"
]

[tool.poetry]
packages = [{include = "src", from = "./"}]

[tool.poetry.dependencies]
python = "^3.13.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
